Move copied `prepare_call_arguments`
